+++ CreationsForge.Bethesda.Assets/Files/AssetFileResolutionStatus.cs
@@ -6,5 +6,6 @@ public enum AssetFileResolutionStatus
     MissingAbsoluteFile,
     MissingDataFolder,
     MissingLooseFile,
+    ResolvedArchiveEntryInMemory,
     ArchiveExtractionUnsupported
 }