+++ CreationsForge.Bethesda.Assets/Resources/BethesdaAssetReadStatus.cs
@@ -1,0 +1,12 @@
+namespace CreationsForge.Bethesda.Assets.Resources;
+
+public enum BethesdaAssetReadStatus
+{
+    ReadLooseFile,
+    ReadArchiveEntry,
+    MissingAbsoluteFile,
+    MissingDataFolder,
+    MissingLooseFile,
+    ArchiveReaderUnavailable,
+    ArchiveEntryMissing
+}