+++ CreationsForge.Bethesda.Assets/Resources/BethesdaAssetSourceType.cs
@@ -1,0 +1,8 @@
+namespace CreationsForge.Bethesda.Assets.Resources;
+
+public enum BethesdaAssetSourceType
+{
+    None,
+    LooseFile,
+    Archive
+}