+++ CreationsForge.Bethesda.Assets/Resources/IBethesdaAssetProvider.cs
@@ -1,0 +1,6 @@
+namespace CreationsForge.Bethesda.Assets.Resources;
+
+public interface IBethesdaAssetProvider
+{
+    BethesdaAssetReadResult TryReadAsset(BethesdaAssetReadRequest request);
+}